Moves Dataset.allocate_ids to connection module.
Addresses fifth part of #477.
